Add decomposition of density matrix into Pauli matrices.
Create logical_decomposition of LogicalDensityMatrix by first constructing a logical representation (2^n_logical_qubits x 2^n_logical_qubits), then decomposing this into a Pauli tensor basis using an algorithm by Hamaguchi et al. based on the fast Walsh-Hadamard transform. Also return a `delta`, which gives a measure of the size of the part of the density matrix not in the code space.

Also fix a minor bug in `logical_decomposition` for the `LogicalStateVector` class when calculating `delta`.